changes in services

+++ global-service/src/main/java/in/cdac/university/globalService/bean/ApplicationStatusBean.java
@@ -15,7 +15,7 @@ public class ApplicationStatusBean {
 
 
     @ComboKey
-    private Long unumApplicationStatusId;
+    private Integer unumApplicationStatusId;
 
 
     private Integer unumIsvalid;
